requirements.txt
networkx==3.4.2
PyYAML
forex-python